+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **39.2.0:**
 
 - Load-By-Lightning is a new feature, which charges up the battery, when struck by a lightning bolt. Closes (#1)
+- Block-Extraction allows the player to charge the battery, by discharging an energetic block. Closes (#2)
 
 **39.1.0:**